+++ korus-ui/components/MaskedInput/handlers.ts
@@ -150,21 +150,18 @@ export const createKeyDownHandler = (
  * @param {MaskedInputProps} params.props - properties of component
  * @param {SetState<InputValueType>} params.setInputValue - set state function for component's masked value
  * @param {SetState<string>} params.setValue - set state function for component's value
- * @param {string} params.value - component's value
  *
- * @returns {() => void} - reset handler
+ * @returns {(value: string) => void} - reset handler
  */
 export const createResetHandler = ({
   props,
   setInputValue,
   setValue,
-  value,
 }: {
   props: MaskedInputProps,
   setInputValue: SetState<InputValueType>,
   setValue: SetState<string>,
-  value: string,
-}) => () => {
+}) => (value: string) => {
   setValue(value);
 
   if (isFunction(props.onChange)) {
@@ -178,5 +175,9 @@ export const createResetHandler = ({
     props.onChange(customEvent);
   }
 
-  setInputValue(maskValue(value, props.mask, props.placeholderChar));
+  if (value) {
+    setInputValue(maskValue(value, props.mask, props.placeholderChar));
+  } else {
+    setInputValue('');
+  }
 };