+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - MaskedInput: Fixed backspace
 - Fixed test coverage report for Windows
+- MaskedInput: Remove mask on reset
 
 
 ## [0.39.0] - 2020-01-26